Add replay command and update list command
Make list show the timestamps in ISO format.

Change database recording of run to jsonify the description,
which is usually a list of command line args.

Add DB methods for querying data written to the database.

Add replay method for retrieving a given run and printing
it in the same format as the monitor command.

Signed-off-by: Doug Hellmann <doug.hellmann@gmail.com>